% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotprod.R
\name{dotprod}
\alias{dotprod}
\title{Calculate dot product for two vectors}
\usage{
dotprod(x, y)
}
\arguments{
\item{x}{A numeric vector}

\item{y}{A numeric vector}
}
\value{
The dot product of \code{x} and \code{y}
}
\description{
Calculate dot product for two vectors
}
\examples{
x <- c(1, 1)
y <- c(2, 3)
dotprod(x, y)
}
